wrap: Add basic sftp support to wrap.get_data()
This requires any credentials to be supplied in the url, or some other
means of authentication (such as an identity file configured for the
user) to be used. Supplying a password in the URL is not supported.

Mainly for testing purposes, "source_identityfile" can be used to point
out an identity file. Also mainly for testing purposes, "source_hostkey"
can be used to specify a trusted public key for the host.